Fix watchdog, fix overlay state, cleanup.
There was a watchdog that attempted to reset the state and disable the overlay, but
that 'watchdog' timer was never 'woken' and thus every minute the watchdog would bark!
This has been fixed, and the actions taken are also improved.

If you have a chartplotter =and= OpenCPN, and the plotter has already activated the radar
then OpenCPN will start drawing the radar image immediately when it is started, there is no
need to press the "radar" button. If you don't want overlay, you can always disable the plugin.

Some general cleanup.