Fix 4G radar range
A new approach was used to determine the 4G radar ranges. As one of our testers reported,
the Navico chartplotters do not increase range when you scroll the chart image. That suddenly
showed me a way to measure what the actual range of the radar data is that is reported: set up
radar and chart side by side, set the radar image to a particular range and then zoom in on the
border of the radar image to see how far away that is.

It was kind of surprising to see that the returned data stream is, on most ranges, up to 100%
further away than the desired range... At first I did not want to believe this, but it is really
true! Just use the above procedure to corroborate.

Once this was done it was a lot easier to fill a spreadsheet and find out what the most likely
correction factor is. Instead of it using a correction factor like sqrt(2) it turned out to be
simple and elegant. As we already know, there are two 16 bit range values, with only
one of them being set at a time. It turns out that the bigger one of the two is exactly
8 bits bigger, i.e. there is a factor of 256 between the big and small ranges.

The overall range turns out to be a very good fit for the value 4, which also means a bit
shift (2 bits per meter). Seems logical to me, so it is probably correct.