Revise (auto)zoom strategy + chart scaling
This fixes two issues reported by Henning.

First of all, the automatic range calculation is now based on how much -would- fit on the
screen if the boat were in the center, plus 50%. This is what the plotters do as well.
This does mean that if you pan more than a bit the radar image border will be seen; previously
we'd zoom out the radar screen in order to attempt to illuminate any objects on the panned
screen.

Secondly, there was a bug that surfaced if you zoomed in and out. I think this is 4G specific,
which would explain why others haven't seen it. The 4G sends out up to 2048 different scan lines.
It looks as if it does not send out all scanlines on all ranges. If you zoomed from 4 km to 2 km
some of the scanlines would not be updated, and the range calculation which was
'select max(range) from scanlines' would still see a range of a scanline that would now no
longer be reported, and thus be stuck at 4 km. The quick fix is to just scale to the current
radar range, i.e. the last one received. This does mean that after a zoom some old scanlines
are zoomed incorrectly on the screen, but this can be fixed. I am not bothering now as this
will be refactored anyhow.