Refactor drawing of radar blobs.
draw_blob_gl() used to draw too big blobs close by. Because it drawed over the same pixel (after translation from
the (alpha,distance) plane to the cartesian plane that we draw onto using OpenGL the transparency seemed to be reduced.

The new code is more careful in what it draws exactly. This seems to have the desired effect:

- Transparency over the entire image is constant.
- Range seems more exact as well.

Resolves #15.
Resolves #14.