Show and hide radar control when toolbar is clicked.
When using a touch screen the right mouse dialog entry to open the radar control is
difficult, and no function should be hidden behind a right mouse click.
Now it is only needed to open it when you have explicitly closed it. Or you click
on the toolbar button twice.