Add packet/spoke feedback on screen.
If you set VerboseLog=1 in the ini file the Advanced Control will show feedback on the # of packets and spokes received.

The data is 'Packets <received>/<broken>' and 'Spokes <received>/<broken>/<missing>'.
Other than the immediate start of the radar the number of broken and missing spokes should be zero.