Fix scan age calculation to be time based.
When you scroll or move the image the rendering code can be called many times a second.
This causes the radar image to be removed as the 'age' calculation depended on the
renderer being called every second. This has now been changed to a date based
algorithm, so it is no longer dependent on that.

Also lock the data structure whilst drawing + changing, this does not seem to make
a huge difference but at least it is safer.