Add alert sound selection to preferences dialog.
Add a way to select and test a sound to be used in the Guard Zone Alerts. If the file is not set then use wxBell() -- which should
be the default sound.

The sound function used is the 1.10 API plugin PlugInPlaySound() so if there is a bug on Linux or Mac with this then it is not the
fault of the plugin but core OpenCPN.

Fix #22.