Rename NMEA0183 class to br_NMEA0183.
Based on a suggestion by @ptulp, this renames NMEA0183 to br_NMEA0183. The better solution would be to hide the class so that the DLL does not load or export it outside the DLL/.so/.dylib, but I don't have the time to find out how to do that on all platforms, in C++ (gosh, it would be so simple for me in Java...)

Remove unneeded NMEA sentences.