Initial commit for v2.0.
This commit contains the largest change since the plugin was started.
The code has been refactored to be more maintainable, easier to understand (I hope!)
and easier to share with possible other radar software.

File that start with br24* are really Navico specific, and must be rewritten for other
brangs. Files that start with other stuff are really only dependent on the concept of
a radar.

This also refactors the dual radar support that was introduced in v1.2, and the
groundwork for dual independent radar only displays.

Code compiles on all three platforms, but is totally untested.

Furthermore:
- Timed transmit is disabled
- Guard zone support is not yet complete.