Add C++ namespace
This remaps all plugin code to reside in the 'br24' namespace.

This makes sure the code in this plugin does not interfere with or is interfered
by code in other plugins or within OpenCPN itself.

This also undoes the earlier rename of NMEA0183 to br_NMEA0183 as this is now no longer needed.

I started this in particular because I noticed that all sorts of plugins now use jsonreader.

Note that this has been done in such a way that it is still simple to share the shader or vertex code with other radars.