Safeguard against OpenGL context mixup
At least when changing color scheme can the glCanvas in OpenGL forget to switch back to its own context.
With this fix we switch in a fake 'zero' or 'empty' OpenGL context so that any delete texture operations
don't happen on our texture. This will cause a GL object leak in OpenGL, but that seems preferable.