Fix crash & recursiveness.
It turns out that we can't play with visibility of the radar from the Render calls in br24radar_pi.
The wxAuiManager calls Update() which calls the OpenCPN ChartRedraw which calls Plugin->Render...
so we can't do any wxAuiManager calls from this.

Fix this by switching to a separate timer that calls br24radar_pi::Notify (= the old DoTick method).