Fix #132 - First phase.
When you click in the main chart menu a lollipop appears in the radar window(s).

It is also shown in the lower left corner as distance and direction.

Note that this ain't a real VRM/EBL (shown as distance and direction) and you can't reset it or click from the radar window.