Fix #132 - Phase 2
There are VRM/EBL placement buttons now but you still can't set the mouse cursor in the radar window.

I also see that the lollipop is just a simple cursor symbol on the HDS, that needs to change as well.

Furthermore, if you have a hidden radar but use the control dialog for the overlay the VRM/EBL button should be hidden.