Fix #163, Fix #167
Work on GuardZone and feedback.

The guard zone display is only shown if the 'top level' menu is shown, so if you are halfway through editing some control it does not interfere.
The guard zone display is hidden and the top level menu shown again if the bogeys have gone away, even if you have not confirmed.

The confirmation lasts a while, so if a bogey leaves the guard zone but then re-enters within 10 seconds you are not alarmed.

The confirm status is shown in the PPI radar window bottom left, where it now says (confirmed) after a bogey count if the confirmation status still holds.
You can see this go away after 10s when the bogeys are no longer in the guard zone.