Fix rotations and headings
Although Douwe didn't have any issues the GL displays were not displaying at the correct rotations for me.

To find what was going on I have also introduced a new Preference setting to change the ARPA edge color, and its opacity (alpha). The default is white, no transparency (fully opaque). Also, the ARPA edges are now drawn on top of the radar echoes to make them even clearer.

For me it now works completely on the radar windows, and in North Up mode on the overlay. Course up on overlay may need a final nudge but that will have to be tested whilst moving.

Furthermore, and that is maybe even the biggest change -- the heading at which radar spokes are stored now comes directly from the radar spoke, and not via an indirect m_hdt value. This stabilized the display enormously (I was testing whilst yawing quite badly today, up to 5 degrees to port and to starboard.) In North and Course Up modes the radar window is now completely stable, as it should be.