Speed up shutdown of receive threads.
We now shut both threads in parallel by calling RadarInfo.Shutdown() which will now remove
the dialogs and send a shutdown message to the receive thread.

There is a TEST_THREAD_RACES preprocessor definition that can be defined in order to
test this. Note that the log files *timestamp* should be believed, not the order in which the records
appear in the logfile. There are separate buffers for every thread, so the logfile is not in the correct order at all times.