Issue #288. Issue #287.
Rework timer operation completely.
The timer is now on the main class, which makes more sense.
When RefreshRate = 1, we don't call any extra timer calls at all.
When RefreshRate > 1 we sync from when OpenCPN calls us, and take
the time it takes to draw into account.

At RefreshRate=5 we now draw 16 times per second, not 21, and the rates
go up exponentially now (1 - 2 - 4 - 8 - 16). Have to see with a live
radar how that looks.

CPU load in VirtualMachines is a lot better now, still have to test with
low end hardware.