ci: cloudsmith upload: Add metadata and tagged repo.
Use separate repos for untagged test builds and tagged releases.
The default repo names are hardcoded, could be overridden using
environment variables CLOUDSMITH_STABLE_REPO  and
CLOUDSMITH_UNSTABLE_REPO.

If a commit is tagged the artifacts are sent to the stable repo,
else the unstable.

Add some metadata to the uploaded artifacts to make the repo
somewhat easier to scan manually.