build: metadata: Add arch suffix to ubuntu target (#19).
Add a -x86_64 suffix to the <target-os> field in the metadata. This
should not be required, but it is.

raspbian and debian is treated in the same way, there are no specific
code paths in opencpn for it. I have no arm hw to test on, though.

See: https://github.com/OpenCPN/OpenCPN/issues/2003