build: Format cmake files using cmake-format.
The cmake files have evolved over time, have ancient roots and
nowadays have inconsistent formatting which makes them harder
to read and overall gives an outdated impression.

To avoid possible discussions, use the automatic formatting of
https://github.com/cheshirekow/cmake_format which is used also in
main opencpn. In particular, the settings file is a verbatim
copy from the main project.

The net result is a modern, consistent formatting which eventually
make the code easier to read after some time to get used to it.