libs: jsoncpp: Update and bundle in separate dir (#33).
The jsoncpp file generates some nasty warnings. The basic
message is that the files are outdated.

Adding external libraries in src/ as actually a bad habit,
making updates and tracking of dependencies harder. Use
the model from main opencpn with a separate libs dir, and
use cmake's interface libraries to handle it. Add a fresh
jsoncpp copy, upstream commit 96e9cda.

And, not least, include the tests.