config: Don't duplicate configuration in C files.
Version and API level info which are available in build configuration
are duplicated as #define constants in the C header file. This is wrong,
these should be defined in the single place CMakeLists.txt.

Add a config.h/config.h.in to make the CMakeLists constants visible in
the C code and drop the duplicated definitions. While on it, clean up
the #include ordering.