Write Navico commands to socket bound to proper interface
We need to receive data on IN_ADDR_ANY, this is known to work.
But to _send_ data we cannot use this socket on Linux, as Linux then
uses the first ethernet card it can lay its hand on.

So the sending to radar is now handled by NavicoControl.cpp, which has
such a socket, using an extra TransmitCmd() method with an extra send
address argument.

This also refactors the SetSendMultiCastAddress() method on RadarControl
and moves it into the specialised NavicoControl and RME120Control
classes, where it belongs as the other radars don't have this.
Furthermore, this is now named SetSendAddress(), whether it is a
multicast address or not is irrelevant for sending.