Add infrastructure for developer team builds
See Makefile, build.sh and build.cmd.

The idea is that build.sh and build.cmd contain the extra variables
and calls necessary to run from a straight shell or cmd prompt on our
build machines.

At the moment this only works on Kees' machines, others still to be
added by themselves.