ci: Make flatpak/raspbain run in unique build dir
The raspbian and flatpak builds does not run in docker but rather
in the local environment. They need to stay in the current build tree
to use the caching (flatpak caches and, for raspbina, the master.zip
download).

In order to avoid conflicst in the generated data, these needs a
separate build directory. Furhermore, git-push need to create the
plugins clone in the build directory for the same reasons.