CMakeLists: Only link libjsoncpp if not done by Plugin.gl (#330).
This should be acceptable as a clean bugfix. It will not add the
wxjson subdirectory twice, the root problem in #330.

The catch is that semantics will change: A plugin which links to
wxJson will fail since wxjson will not be visible until all plugins
are linked.

However, I'm not aware of any plugin which depends on wxJson. And
if there indeed is such a plugin, it can be handled by Plugin.gl
by adding wxjson before. This is a lesser eveil than the current #330
situation

Closes: #330