appveyor: Use the regular win_deps.bat script for builds.
Using the win_deps.bat script in the CI builds ensures that possible
problems are resolved quickly. This makes the windows builds work
as Linux and MacOS which also uses the same script for both
manual and CI builds.

Add the no-check-certificate to wget -- seems that the clock in the
appveyor build box isnt set, causing certificate validation failures.

Clean up the yaml syntax and make appveyor.yml pass yamllint.