win_deps.bat: Make PATH handling more consistent.
Use the pathman tool to set PATH in the registry rather than the
current session. Besides that the PATH set now is persistent
pathman also ensures that path elements are not duplicated.

There seems to be some changes in how Windows handles PATH
lately, not sure. The fixes are hopefully sound anyway.