Flatpak: configure the yaml manifest
Configure a new manifest file in the build directory as a copy of the
yaml manifest in the flatpak/ directory and use that.

TL;DR:
- Existing manifests works as-is without changes.
- Existing manifests can add @include to use snippets in opencpn-libs,
  see below.
- New plugins gets a generic manifest which can be used without changes
  if there are no dependencies.

During configuration, @plugin_name is replaced with PLUGIN_API_NAME
from Plugin.cmake. This makes it possible to use a standard manifest
in plugins without dependencies. @plugin_name is expected to be used in
the id: line in the manifest initial part.

@app_id is replaced with the base part of the id: line, possibly after
substituting @plugin_name. Tha part after last dot is used. @app_id is
expected to be used in CMAKE_INSTALL_PREFIX.

The original manifest in flatpak/ can contain lines like:

  - @include libs/glu.yaml

Such a line is replaced by the contents of the file mentioned after
@include, here 'libs/glu.yaml'. The path is relative to the project
top directory. The indentation of the '-' character is added to each
line in the included file.

The @include mechanism  makes it possible to use shared library
snippets which are easier to maintain.

The Shipdriver manifest contains examples on all of these mechanisms.