Fix #201: HALO mode support
Move the low range of radar returns (nibble values 1 and 2) into
the 'blue' range so that they are visible.

Add a (common to all radars) Threshold button so you can select
whether you want to see this.

Note that is not a solution for Garmin xHD and Raymarine HD which
have a full 8 bit per pixel, but those don't have bird mode, and
we never received any comments from people running that radar that
they are missing low intensity returns.