#201: Bird mode/threshold improvements
Minimal limit for blue is now what the radar independent part can support
which is 33 (because of the 32 history values used). This is now also
the default.

Make threshold's 10% step values start at 33 so that we see a difference
from 10% and up.

Add a ring that shows the different colours in the emulator.